+++ test/unit/internal/rekor/test_client.py
@@ -21,24 +21,32 @@ from sigstore.transparency import LogInclusionProof
 
 class TestRekorInclusionProof:
     def test_valid(self):
-        proof = LogInclusionProof(log_index=1, root_hash="abcd", tree_size=2, hashes=[])
+        proof = LogInclusionProof(
+            log_index=1, root_hash="abcd", tree_size=2, hashes=[], checkpoint=""
+        )
         assert proof is not None
 
     def test_negative_log_index(self):
         with pytest.raises(
             ValidationError, match="Inclusion proof has invalid log index"
         ):
-            LogInclusionProof(log_index=-1, root_hash="abcd", tree_size=2, hashes=[])
+            LogInclusionProof(
+                log_index=-1, root_hash="abcd", tree_size=2, hashes=[], checkpoint=""
+            )
 
     def test_negative_tree_size(self):
         with pytest.raises(
             ValidationError, match="Inclusion proof has invalid tree size"
         ):
-            LogInclusionProof(log_index=1, root_hash="abcd", tree_size=-1, hashes=[])
+            LogInclusionProof(
+                log_index=1, root_hash="abcd", tree_size=-1, hashes=[], checkpoint=""
+            )
 
     def test_log_index_outside_tree_size(self):
         with pytest.raises(
             ValidationError,
             match="Inclusion proof has log index greater than or equal to tree size",
         ):
-            LogInclusionProof(log_index=2, root_hash="abcd", tree_size=1, hashes=[])
+            LogInclusionProof(
+                log_index=2, root_hash="abcd", tree_size=1, hashes=[], checkpoint=""
+            )